requirements.txt
mcp-proxy-adapter>=1.0.0
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
GitPython>=3.1.0 